Included bullet point details in description
Bullet points were run through google translate then revised by hand.

Translator Notes:
The word I translated to "Friends" is 童鞋, which literally translates to "children's shoes". This is a slang term that affectionately refers to a fellow young person and sounds like the word 同学 (classmates). I chose "Friends" because it seemed the closest to the feeling the author was intending.